services and discount select api

+++ controllers/organizer/events/categories.js
@@ -1,4 +1,5 @@
 const categoryModel = require('../../../models/categories.model');
+const organizerModel = require('../../../models/organizers.model');
 const responseManager = require('../../../utilities/response.manager');
 const mongoConnection = require('../../../utilities/connections');
 const constants = require('../../../utilities/constants');
@@ -6,32 +7,38 @@ const mongoose = require('mongoose');
 exports.addcategory = async (req, res) => {
     res.setHeader('Access-Control-Allow-Origin', '*');
     if (req.token.organizerid && mongoose.Types.ObjectId.isValid(req.token.organizerid)) {
-        const { categoryid, category_name } = req.body;
         let primary = mongoConnection.useDb(constants.DEFAULT_DB);
-        if (categoryid && categoryid != '' && mongoose.Types.ObjectId.isValid(categoryid)) {
-            if (category_name && category_name.trim() != '') {
-                let obj = {
-                    category_name: category_name,
-                    updatedBy: mongoose.Types.ObjectId(req.token.organizerid)
-                };
-                await primary.model(constants.MODELS.categories, categoryModel).findByIdAndUpdate(categoryid, obj);
-                let categoryData = await primary.model(constants.MODELS.categories, categoryModel).findById(categoryid);
-                return responseManager.onSuccess('Organizer event category data updated successfully!', categoryData, res);
+        let organizerData = await primary.model(constants.MODELS.organizers, organizerModel).findById(req.token.organizerid).select('-password').lean();
+        if (organizerData && organizerData.status == true && organizerData.mobileverified == true) {
+            const { categoryid, category_name } = req.body;
+            let primary = mongoConnection.useDb(constants.DEFAULT_DB);
+            if (categoryid && categoryid != '' && mongoose.Types.ObjectId.isValid(categoryid)) {
+                if (category_name && category_name.trim() != '') {
+                    let obj = {
+                        category_name: category_name,
+                        updatedBy: mongoose.Types.ObjectId(req.token.organizerid)
+                    };
+                    await primary.model(constants.MODELS.categories, categoryModel).findByIdAndUpdate(categoryid, obj);
+                    let categoryData = await primary.model(constants.MODELS.categories, categoryModel).findById(categoryid);
+                    return responseManager.onSuccess('Organizer event category data updated successfully!', categoryData, res);
+                } else {
+                    return responseManager.badrequest({ message: 'Invalid add category name can not be empty, please try again' }, res);
+                }
             } else {
-                return responseManager.badrequest({ message: 'Invalid add category name can not be empty, please try again' }, res);
+                if (category_name && category_name.trim() != '') {
+                    let obj = {
+                        category_name: category_name,
+                        createdBy: mongoose.Types.ObjectId(req.token.organizerid),
+                        updatedBy: mongoose.Types.ObjectId(req.token.organizerid)
+                    };
+                    let categoryData = await primary.model(constants.MODELS.categories, categoryModel).create(obj);
+                    return responseManager.onSuccess('Organizer event category created successfully!', categoryData, res);
+                } else {
+                    return responseManager.badrequest({ message: 'Invalid add category name can not be empty, please try again' }, res);
+                }
             }
         } else {
-            if (category_name && category_name.trim() != '') {
-                let obj = {
-                    category_name: category_name,
-                    createdBy: mongoose.Types.ObjectId(req.token.organizerid),
-                    updatedBy: mongoose.Types.ObjectId(req.token.organizerid)
-                };
-                let categoryData = await primary.model(constants.MODELS.categories, categoryModel).create(obj);
-                return responseManager.onSuccess('Organizer event category created successfully!', categoryData, res);
-            } else {
-                return responseManager.badrequest({ message: 'Invalid add category name can not be empty, please try again' }, res);
-            }
+            return responseManager.badrequest({ message: 'Invalid organizerid to update event, please try again' }, res);
         }
     } else {
         return responseManager.badrequest({ message: 'Invalid token to create event category data, please try again' }, res);
@@ -41,11 +48,17 @@ exports.listcategory = async (req, res) => {
     res.setHeader('Access-Control-Allow-Origin', '*');
     if (req.token.organizerid && mongoose.Types.ObjectId.isValid(req.token.organizerid)) {
         let primary = mongoConnection.useDb(constants.DEFAULT_DB);
-        primary.model(constants.MODELS.categories, categoryModel).find({ createdBy: mongoose.Types.ObjectId(req.token.organizerid) }).lean().then((categories) => {
-            return responseManager.onSuccess('Categories list!', categories, res);
-        }).catch((error) => {
-            return responseManager.onError(error, res);
-        })
+        let organizerData = await primary.model(constants.MODELS.organizers, organizerModel).findById(req.token.organizerid).select('-password').lean();
+        if (organizerData && organizerData.status == true && organizerData.mobileverified == true) {
+            let primary = mongoConnection.useDb(constants.DEFAULT_DB);
+            primary.model(constants.MODELS.categories, categoryModel).find({ createdBy: mongoose.Types.ObjectId(req.token.organizerid) }).lean().then((categories) => {
+                return responseManager.onSuccess('Categories list!', categories, res);
+            }).catch((error) => {
+                return responseManager.onError(error, res);
+            })
+        } else {
+            return responseManager.badrequest({ message: 'Invalid organizerid to update event, please try again' }, res);
+        }
     } else {
         return responseManager.badrequest({ message: 'Invalid token to get category data, please try again' }, res);
     }
@@ -53,13 +66,19 @@ exports.listcategory = async (req, res) => {
 exports.getonecategory = async (req, res) => {
     res.setHeader('Access-Control-Allow-Origin', '*');
     if (req.token.organizerid && mongoose.Types.ObjectId.isValid(req.token.organizerid)) {
-        const { categoryid } = req.body;
         let primary = mongoConnection.useDb(constants.DEFAULT_DB);
-        if (categoryid && categoryid != '' && mongoose.Types.ObjectId.isValid(categoryid)) {
-            let categoryData = await primary.model(constants.MODELS.categories, categoryModel).findById(categoryid);
-            return responseManager.onSuccess('Categories data !', categoryData, res);
+        let organizerData = await primary.model(constants.MODELS.organizers, organizerModel).findById(req.token.organizerid).select('-password').lean();
+        if (organizerData && organizerData.status == true && organizerData.mobileverified == true) {
+            const { categoryid } = req.body;
+            let primary = mongoConnection.useDb(constants.DEFAULT_DB);
+            if (categoryid && categoryid != '' && mongoose.Types.ObjectId.isValid(categoryid)) {
+                let categoryData = await primary.model(constants.MODELS.categories, categoryModel).findById(categoryid);
+                return responseManager.onSuccess('Categories data !', categoryData, res);
+            } else {
+                return responseManager.badrequest({ message: 'Invalid category id to get item data, please try again' }, res);
+            }
         } else {
-            return responseManager.badrequest({ message: 'Invalid category id to get item data, please try again' }, res);
+            return responseManager.badrequest({ message: 'Invalid organizerid to update event, please try again' }, res);
         }
     } else {
         return responseManager.badrequest({ message: 'Invalid token to get category, please try again' }, res);
@@ -68,13 +87,19 @@ exports.getonecategory = async (req, res) => {
 exports.removecategory = async (req, res) => {
     res.setHeader('Access-Control-Allow-Origin', '*');
     if (req.token.organizerid && mongoose.Types.ObjectId.isValid(req.token.organizerid)) {
-        const { categoryid } = req.body;
         let primary = mongoConnection.useDb(constants.DEFAULT_DB);
-        if (categoryid && categoryid != '' && mongoose.Types.ObjectId.isValid(categoryid)) {
-            await primary.model(constants.MODELS.categories, categoryModel).findOneAndRemove(categoryid);
-            return responseManager.onSuccess('Category removed sucecssfully!', 1, res);
+        let organizerData = await primary.model(constants.MODELS.organizers, organizerModel).findById(req.token.organizerid).select('-password').lean();
+        if (organizerData && organizerData.status == true && organizerData.mobileverified == true) {
+            const { categoryid } = req.body;
+            let primary = mongoConnection.useDb(constants.DEFAULT_DB);
+            if (categoryid && categoryid != '' && mongoose.Types.ObjectId.isValid(categoryid)) {
+                await primary.model(constants.MODELS.categories, categoryModel).findOneAndRemove(categoryid);
+                return responseManager.onSuccess('Category removed sucecssfully!', 1, res);
+            } else {
+                return responseManager.badrequest({ message: 'Invalid category id to get item data, please try again' }, res);
+            }
         } else {
-            return responseManager.badrequest({ message: 'Invalid category id to get item data, please try again' }, res);
+            return responseManager.badrequest({ message: 'Invalid organizerid to update event, please try again' }, res);
         }
     } else {
         return responseManager.badrequest({ message: 'Invalid token to get category data, please try again' }, res);
